Add pagemap array to tcb struct.
This will be useful for correctly freeing pages and for cloning address
spaces in fork().
