+++ kernel/threads.h
@@ -3,6 +3,7 @@
 
 #include "list.h"
 #include "page.h"
+#include "pagemap.h"
 
 #include <stdint.h>
 #include <stddef.h>
@@ -34,6 +35,7 @@ typedef struct {
   size_t text_length;
   enum fpu_state fpu_state;
   char (*fpu_buf)[512];
+  pagemap pm;
 } tcb;
 
 extern tcb *running_tcb;
